Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 int main(){
 const double PI = 3.14;
-int r;
-scanf("%d" ,&r);
+double r;
+scanf("%lf" ,&r);
 double area = PI*r*r;
 printf("Area: %.2lf",area);
 return 0;
